tests/requirements.txt
pytest-django==4.5.2
pytest-xdist==3.2.1
